Halve memory usage.
There is no good reason to load the data for both the txt2img and
img2img pipelines, because the data is the same. There's even less
good reason to keep it duplicated in memory (about 2 GiB wasted) and
shuffle them between CPU and CUDA.

To debug this, add a function to find and report duplicate pytorch
tensors in memory.